DT: Make compressor geometry its own library
