+++ data_types/BUILD
@@ -29,6 +29,7 @@ py_library(
     srcs = ["centrifugal_compressor.py"],
     visibility = ["//ccpd:__subpackages__"],
     deps = [
+        ":centrifugal_compressor_geometry",
         ":thermo_point",
         ":three_dimensional_blade",
     ],
@@ -37,6 +38,16 @@ py_library(
 py_library(
     name = "three_dimensional_blade",
     srcs = ["three_dimensional_blade.py"],
+    visibility = ["//ccpd:__subpackages__"],
+    deps = [
+        ":centrifugal_compressor_geometry",
+        "@python_deps_numpy//:pkg",
+    ],
+)
+
+py_library(
+    name = "centrifugal_compressor_geometry",
+    srcs = ["centrifugal_compressor_geometry.py"],
     visibility = ["//ccpd:__subpackages__"],
     deps = ["@python_deps_numpy//:pkg"],
 )
